+++ spec/integration/resource/api600/c7000/logical_interconnect_group/delete_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::C7000::LogicalInterconnectGroup
+klass = OneviewSDK::API600::C7000::LogicalInterconnectGroup
 RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
-  let(:current_client) { $client_500 }
-  include_examples 'LIGC7000DeleteExample', 'integration api500 context'
+  let(:current_client) { $client_600 }
+  include_examples 'LIGC7000DeleteExample', 'integration api600 context'
 end