+++ spec/unit/resource/api600/synergy/logical_interconnect_group.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-RSpec.describe OneviewSDK::API500::Synergy::LogicalInterconnectGroup do
+RSpec.describe OneviewSDK::API600::Synergy::LogicalInterconnectGroup do
   include_context 'shared context'
 
-  it 'inherits from OneviewSDK::API300::Synergy::LogicalInterconnectGroup' do
-    expect(described_class).to be < OneviewSDK::API300::Synergy::LogicalInterconnectGroup
+  it 'inherits from OneviewSDK::API600::C7000::LogicalInterconnectGroup' do
+    expect(described_class).to be < OneviewSDK::API600::C7000::LogicalInterconnectGroup
   end
 end